+++ react/typings/gtm.d.ts
@@ -1,5 +1,6 @@
 export interface AnalyticsEcommerceProduct {
   id: string
+  skuId: string
   name: string
   category: string
   brand: string