+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Renamed `ecommerce` event property to `ecommerceV2` to match checkout's property
+- All events are now using the productId as the main product identifier
+- Product name doesn't include the variation anymore
+
+### Added
+- SKU id field to all events that contains product ids
+- `list` property to `productView` (`productDetail`) event
+- `position` property to `productClick` event
+
+### Fixed
+- `productClick` now considers the default seller when getting the item's price
 
 ## [2.9.2] - 2021-03-30
 ### Fixed